#74 factorized asynchronous queue system into tools/asynchronous_client
